Support JS SDK v9 (#642)
