--- web-start/functions/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "friendlychat-codelab",
-  "description": "Firebase SDK for Cloud Functions codelab",
-  "dependencies": {
-    "firebase-admin": "~9.9.0",
-    "firebase-functions": "^3.14.1"
-  },
-  "devDependencies": {
-    "eslint": "^4.12.0",
-    "eslint-plugin-promise": "^3.6.0"
-  },
-  "engines": {
-    "node": "12"
-  },
-  "private": true
-}
