+++ web-start/public/index.html
@@ -110,17 +110,6 @@
   </main>
 </div>
 
-<!-- Import and configure the Firebase SDK -->
-<!-- These scripts are made available when the app is served or deployed on Firebase Hosting -->
-<!-- If you do not want to serve/host your project using Firebase Hosting see https://firebase.google.com/docs/web/setup -->
-<script src="/__/firebase/8.6.7/firebase-app.js"></script>
-<script src="/__/firebase/8.6.7/firebase-auth.js"></script>
-<script src="/__/firebase/8.6.7/firebase-storage.js"></script>
-<script src="/__/firebase/8.6.7/firebase-messaging.js"></script>
-<script src="/__/firebase/8.6.7/firebase-firestore.js"></script>
-<script src="/__/firebase/8.6.7/firebase-performance.js"></script>
-<script src="/__/firebase/init.js"></script>
-
 <script src="scripts/main.js"></script>
 </body>
 </html>
