+++ web-start/src/firebase-messaging-sw.js
@@ -1,0 +1,1 @@
+// TODO: Set up Firebase Cloud Messaging service worker
