+++ web/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "friendlychat",
+  "version": "1.0.0",
+  "description": "[Firebase Web Codelab](https://firebase.google.com/codelabs/firebase-web)",
+  "private": true,
+  "scripts": {
+    "build": "webpack",
+    "start": "webpack --watch --progress",
+    "visualize-bundle": "webpack --analyze"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "webpack": "^5.48.0",
+    "webpack-bundle-analyzer": "^4.4.2",
+    "webpack-cli": "^4.7.2"
+  },
+  "dependencies": {
+    "firebase": "^9.0.0"
+  }
+}
