+++ web/src/firebase-messaging-sw.js
@@ -1,0 +1,8 @@
+// Import and configure the Firebase SDK
+import { initializeApp } from 'firebase/app';
+import { getMessaging } from 'firebase/messaging/sw';
+import { getFirebaseConfig } from './firebase-config';
+
+const firebaseApp = initializeApp(getFirebaseConfig());
+getMessaging(firebaseApp);
+console.info('Firebase messaging service worker is set up');
